feat:include routes to access redeemHistory and filteredProducts independently
